add providers for swap path
